Añadida interfaz IEncounter. Movido Encounter a library/encounters. Optimizaciones en Encounter -> negaciones son mas lentas que aserciones, el constructor puede estas mas resumido y no es necesario hacer dos for loops, ademas tiraba error por añadir elementos a listas no inicializadas, clonadas las listas de heroes y villanos al momento de removes personajes muertos -> eliminar items de una lista mientras se recorre creo que puede tirar ConcurrentModificacionException o algo asi, en la letra dice que el primer villano debe atacar al primer heroe no al ultimo, ademas habia un villano que no llegaba atacaba a un heroe (ver el else), movido el if heroes < 0 o villains < 0 al principio para molestar nomas. Es mas largo el mensaje que la cantidad de cambios que hice.
