+++ src/Library/encounters/IEncounter.cs
@@ -1,0 +1,7 @@
+namespace RoleplayGame
+{
+    public interface IEncounter
+    {
+        void DoEncounter();
+    }
+}
